backend/ml_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
tensorflow>=2.20.0
numpy>=1.24.0
Pillow>=10.0.0
pymongo>=4.6.0

